+++ Lykke.B2c2Client/Models/Rest/TradeRequest.cs
@@ -15,11 +15,11 @@ namespace Lykke.B2c2Client.Models.Rest
         public Side Side { get; set; }
 
         [JsonProperty("price")]
-        public decimal Price { get; set; }
+        public double Price { get; set; }
 
         /// Quantity in base currency (maximum 4 decimals).
         [JsonProperty("quantity")]
-        public decimal Quantity { get; set; }
+        public double Quantity { get; set; }
 
         public TradeRequest()
         {
